Add oceanic edges
After working hard to make sure that the noise lines up on along
the seam, there's now water along that seam. The height noise on
the ocean floor does still align though, so it's not all for
naught.

The reason for doing this is that a straight line is drawn along the
180th meridian on the vanilla EU4 map. I'm not in a position to assume
that this isn't an engine limitation, and if I have to do this then
it's good to have in place before working on the full image output.