Add full size render output
While image output has always been sort of possible by way of
screenshots, it's not the prettiest way. Screenshots also put a limit
on the resolution supported. But finally, we can render images at
arbitrary resolutions.

A number of things have been added because of, or to aid in this:
- A new uniform, renderOffset.
- The image is flipped vertically, so as to match the screen output.
- An error is raised if the image size requested is too large. This
  limit can be raised later, but 8192*8192 should be plenty.
- An extra octave to the noise, because now we have a higher
  resolution.