Add wedge to height values for reproducibility
Wedge, like the opposite of clamp. It seems that some floating point
height values below 0.5 and some above both round to the same 8-bit
value. Not sure how, it might be some sort of render optimisation.
But it has the effect of making it so that the CPU-side integer
coastline doesn't line up with the GPU-side floating point coastline.

By driving a wedge of sorts around the 0.5 range, we can assure that
the respective sides of 0.5 don't round to the same value and that
the coastline remains the same later in the pipeline.