Make render to texture process simpler
The previous process was necessary due to undefined behaviour
causing crashes with large textures. With that solved, we can just
request the full texture size and render to it, no looping over
texture blocks.

In addition to the fact that undefined behaviour is volatile and
thus undesirable, and this process being quite frankly, simpler,
the new approach is also going to make province generation
significantly easier.