Add basic province generation
This is a massive overhaul. Maybe a bit too large, in hindsight.
But we have provinces! They don't take height differences into
account and they're sort of unevenly packed, but it's a big step.

With this change, it exports province boundaries rather than terrain,
which is a sort of feature regression. A few lines have been commented
out, but they're sort of... on hiatus. They'll come back.