Fix reference to defunct terrain sampler
This is its own commit because I wanted the last commit to be a pure
rename commit as far as the heightmap shader files were concerned. The
sampler might not need to be there at all but I'm just one person with
one computer and I've yet to find a good linter.