Fix platform dependent hash behaviour
This is a breaking change, but since we're pre-alpha, it doesn't
really matter.

The hash function that came with iq's gradient noise code produces
different output depending on hardware and potentially other factors
like driver behaviour.

For this reason, it's been necessary to replace it with an integer
based hash function which produces identical behaviour, at least from
the hash. Differences in rounding behaviour may lead to very slightly
different output after that though. But that's out of scope for now...